feat: implement VRF attachment operations for VLAN management
